Added media queries + minor changes
- On line 91, moved "@include transition" out of &:hover
because the transition was not working when hover effect ended.
Now it works for both transitions

- On line 53, reduced bottom padding  (commented out)
On my phone, button was getting cut off at an awkward place

- changed font color to lighter green
it's just my own preference, so please feel free to change it back!

- added media query to add padding to top of page on larger screens

- added media query for landscape mobile phones
This one was because the background IMG was almost entirely cut off
when I tried landscape mode on my phone. I wasn't sure of a good fix,
I ended up cropping the background img so I could position it
differently, but I think this still needs adjustments.

- changed background color to match with image background
I was seeing flickers of the other background color when scrolling,
which was distracting.   Also it helps with the cropped version of the
picture.